pyproject.toml
[project]
name = "ocr-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.128.0",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
]
